Função de Listar usuários/Habilitar e Desabilitar concluída!
