+++ src/components/Table/Styles.module.scss
@@ -1,0 +1,51 @@
+.container {
+    width: 80vw;
+    padding: 10px;
+
+    table {
+        color: white; 
+        width: 100%;
+        border-collapse: collapse;
+    }
+
+    thead {
+        background-color: #52557D;
+    }
+
+    text-align: center;
+    vertical-align: middle;
+
+    th {
+        padding: 8px;
+        text-align: left;
+        border-bottom: 1px solid #ddd;
+        text-align: center;
+    }
+
+    td {
+        padding: 8px;
+        text-align: left;
+        border-bottom: 1px solid #ddd;
+        text-align: center;
+    }
+
+    tbody {
+        border-collapse: collapse;
+    }
+
+    tr:nth-child(even) {
+        background-color: #12131b;
+    }
+
+    button {
+        border-radius: 5px;
+        width: 60px;
+        height: 30px;
+        background-color: green;
+    }
+}
+
+.imagem {
+    width: 30px;
+    height: 30px;
+}
