+++ src/services/api.js
@@ -138,10 +138,10 @@ export const ApiCEPList = async (cep = '00000000') => {
   // Retornar apenas a lista
 };
 
-export const ApiUserStatus = async (cpf, statusUser) => {
+export const ApiUserStatus = async (id, statusUser) => {
   let data;
   const object = JSON.stringify({
-    userCpf: cpf,
+    userId: id,
     userStatus: statusUser,
   })
   await fetch(`${url}/backoffice/user/update/status`, {
@@ -152,7 +152,7 @@ export const ApiUserStatus = async (cpf, statusUser) => {
     body: object,
   })
     .then(async (response) => {
-      data = await response.json();
+      data = await response.status;
     })
     .catch(async (error) => {
       console.log(object)
